Add typer, more gvi_score
- Add Typer code to execute from cmd
- Move get_gvi_score inside assign_gvi_to_points.py
